Update default value of `username_template`
